# api/openapi.yaml
openapi: 3.0.3
info:
  title: Orbit API
  version: 0.0.1
  description: Orbit v0.0.1 single-cluster preview API
servers:
  - url: /
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
        expiresIn:
          type: integer
      required: [accessToken, tokenType, expiresIn]
...
    EvidenceGenerateRequest:
      type: object
      properties:
        scopeType:
          type: string
          enum: [namespace, pod]
        namespace:
          type: string
        name:
          type: string
        persist:
          type: boolean
      required: [scopeType, namespace]
security:
  - BearerAuth: []
paths:
...
  /api/v1/auth/login:
    post:
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/inventory/resources:
    get:
      parameters:
        - in: query
          name: kind
          schema: { type: string }
        - in: query
          name: namespace
          schema: { type: string }
        - in: query
          name: name
          schema: { type: string }
      responses:
        '200':
          description: Resource inventory
  /api/v1/findings:
    get:
      parameters:
        - in: query
          name: severity
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: namespace
          schema: { type: string }
        - in: query
          name: kind
          schema: { type: string }
      responses:
        '200':
          description: Findings list
...
  /api/v1/evidence-packs/generate:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvidenceGenerateRequest'
      responses:
        '200':
          description: Generated evidence pack
  /api/v1/evidence-packs:
    get:
      parameters:
        - in: query
          name: scopeType
          schema: { type: string }
        - in: query
          name: namespace
          schema: { type: string }
        - in: query
          name: name
          schema: { type: string }
        - in: query
          name: findingId
          schema: { type: string }
      responses:
        '200':
          description: Evidence pack list
...